Separate readdata from main
